requirements.txt
fastapi
uvicorn
requests
python-dotenv
scipy
weaviate-client